#us client satisfaction: corrected all bugs and integrated client list

+++ android-crmtab/app/src/main/java/fr/pds/isintheair/crmtab/mbalabascarin/uc/edit/crv/view/CrvMainActivity.java
@@ -28,7 +28,7 @@ public class CrvMainActivity extends AppCompatActivity {
 
     Client client;
     Report report;
-    int clientId;
+    long clientId;
     ListView reportListView;
     ReportAdapter adapter;
     private List<Report> reportList = new ArrayList<Report>();
